Reorganized into separate files for spectrometers.
